requirements.txt
aiohttp==3.6.3
astroid==2.4.2
async-timeout==3.0.1
asyncio==3.4.3
attrs==20.3.0
autopep8==1.5.4
beautifulsoup4==4.9.3
certifi==2020.6.20
chardet==3.0.4
deepdiff==5.0.2
discord.py==1.5.1
idna==2.10
importlib-metadata==2.0.0
isort==5.6.4
jsonpickle==1.4.1
lazy-object-proxy==1.4.3
lxml==4.6.1
mccabe==0.6.1
mergedeep==1.3.0
mmh3==2.5.1
multidict==4.7.6
numpy==1.19.3
ordered-set==4.0.2
pep8==1.7.1
psutil==5.7.3
pycodestyle==2.6.0
pylint==2.6.0
PySocks==1.7.1
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2020.4
requests==2.24.0
six==1.15.0
soupsieve==2.0.1
timeago==1.0.13
toml==0.10.2
ujson==4.0.1
urllib3==1.25.11
win32-setctime==1.0.3
wrapt==1.12.1
yarl==1.1.1
zipp==3.4.0
